hw_2/requirements.txt
# Core ML/DL
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.24.0
pandas>=1.5.0
scikit-learn>=1.2.0

# NLP
transformers>=4.25.0
tokenizers>=0.13.0

# Computer Vision
opencv-python>=4.7.0
Pillow>=9.4.0

# Data
requests>=2.28.0
beautifulsoup4>=4.11.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Utilities
tqdm>=4.64.0
python-dotenv>=0.21.0

# Optional: Model explainability
shap>=0.42.0
lime>=0.2.0

# Optional: Optimization
onnx>=1.13.0
onnxruntime>=1.14.0

# Development
jupyter>=1.0.0
ipython>=8.10.0
